Move the writer classes to sink package
