+++ flink/src/main/java/org/apache/iceberg/flink/sink/IcebergStreamWriter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iceberg.flink;
+package org.apache.iceberg.flink.sink;
 
 import java.io.IOException;
 import org.apache.flink.streaming.api.operators.AbstractStreamOperator;
