+++ flink/src/main/java/org/apache/iceberg/flink/sink/PartitionedFanoutWriter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iceberg.flink;
+package org.apache.iceberg.flink.sink;
 
 import java.io.IOException;
 import java.util.Map;
